更新
1、将原先的三个难度中的Boss血量改为有限值，同时加入无尽模式
2、添加了退出按钮，现在可以随时退出游戏
3、优化了dot的显示方式
4、修复了一局游戏结束之后，重新开始游戏，”献祭“的增伤没有重置的bug
5、修复了手机端无法正常滚动dot栏的bug
6、修复了部分dot被偷走后，目标没有随之改变的bug
7、将”七宗罪“的”色欲“改为只延长非”七宗罪“dot，避免了多个”七宗罪“叠加导致的全自动效果
8、修复了敌人动画在被加速时播放出现的bug

+++ Assets/Scripts/Card/CardDataBase.cs
@@ -11,8 +11,8 @@ public static class CardDatabase
     {
         public int id;
         public string name;
-        public int cost;
-        public int value;
+        public ulong cost;
+        public ulong value;
         public int duration;
         public string effect;
         public string imagePath;
@@ -153,12 +153,12 @@ public static class CardDatabase
                     Debug.LogWarning($"卡牌ID解析失败: {fields[0]}");
                     continue;
                 }
-                if (!int.TryParse(fields[2], out int costValue))
+                if (!ulong.TryParse(fields[2], out ulong costValue))
                 {
                     Debug.LogWarning($"卡牌成本解析失败，ID={idValue}: {fields[2]}");
                     continue;
                 }
-                if (!int.TryParse(fields[3], out int valueValue))
+                if (!ulong.TryParse(fields[3], out ulong valueValue))
                 {
                     Debug.LogWarning($"卡牌价值解析失败，ID={idValue}: {fields[3]}");
                     continue;
